requirements.txt
numpy==1.24.3
pandas==2.0.3
tensorflow-macos==2.13.0
tensorflow-metal==1.0.0
scikit-learn==1.3.0
matplotlib==3.7.2
scipy==1.11.2 